Switch to refs, share toolbar
