+++ src/stories/MakeCodeFrame.stories.tsx
@@ -9,6 +9,7 @@ import {
 import { Project } from '../vanilla/pxt.js';
 import { controllerId } from './config.js';
 import StoryWrapper from './StoryWrapper.js';
+import MakeCodeToolbar from './MakeCodeToolbar.js';
 
 const meta: Meta<typeof MakeCodeFrame> = {
   title: 'stories/React/MakeCodeFrame',
@@ -26,16 +27,6 @@ const meta: Meta<typeof MakeCodeFrame> = {
 export default meta;
 type Story = StoryObj<typeof MakeCodeFrame>;
 
-const EditorWithToolbarStyles = {
-  actions: {
-    fontFamily: 'sans-serif',
-    display: 'flex',
-    flexWrap: 'wrap',
-    gap: '5px',
-    margin: '10px 0',
-  } as const,
-};
-
 const MakeCodeEditorWithControls = (
   props: Omit<MakeCodeFrameProps, 'initialProjects'>
 ) => {
@@ -50,192 +41,7 @@ const MakeCodeEditorWithControls = (
   }, []);
   return (
     <>
-      <div style={{ display: 'flex', flexDirection: 'column' }}>
-        <div style={EditorWithToolbarStyles.actions}>
-          <button onClick={() => ref.current!.switchJavascript()}>
-            Javascript
-          </button>
-          <button onClick={() => ref.current!.switchBlocks()}>Blocks</button>
-          <button
-            onClick={async () => {
-              const info = await ref.current!.info();
-              console.log(info);
-            }}
-          >
-            Info
-          </button>
-          <button onClick={() => ref.current!.newProject()}>New project</button>
-          <button
-            onClick={() =>
-              ref.current!.startActivity({
-                activityType: 'tutorial',
-                path: 'microbit-foundation/makecode-tutorials/first-lessons/step-counter',
-              })
-            }
-          >
-            Load tutorial from GitHub
-          </button>
-          <button onClick={() => ref.current!.print()}>Print</button>
-          <button onClick={() => ref.current!.pair()}>Pair</button>
-          <button onClick={() => ref.current!.compile()}>Compile</button>
-          <button onClick={() => ref.current!.saveProject()}>
-            Save project
-          </button>
-          <button onClick={() => ref.current!.unloadProject()}>
-            Unload project
-          </button>
-          <button
-            onClick={() =>
-              ref.current!.openHeader(
-                [...savedProjects.current.values()][0].header!.id
-              )
-            }
-          >
-            Open header
-          </button>
-          <button
-            onClick={() =>
-              ref.current!.importProject({
-                project: defaultMakeCodeProject,
-              })
-            }
-          >
-            Import project (no header)
-          </button>
-          <button
-            onClick={async () => {
-              const result = await ref.current!.shareProject({
-                headerId: [...savedProjects.current!.values()][0].header!.id,
-                projectName: 'Example project name',
-              });
-              console.log(result);
-            }}
-          >
-            Share project
-          </button>
-          <button
-            onClick={() =>
-              ref.current!.setLanguageRestriction('javascript-only' as const)
-            }
-          >
-            Set language restriction
-          </button>
-          <button
-            onClick={async () => {
-              const result = await ref.current!.getToolboxCategories({
-                advanced: true,
-              });
-              console.log(result);
-            }}
-          >
-            Get toolbox categories
-          </button>
-          <button onClick={() => ref.current!.toggleDebugSloMo()}>
-            Toggle debug slow mo
-          </button>
-          <button onClick={() => ref.current!.toggleGreenScreen()}>
-            Green screen
-          </button>
-          <button onClick={() => ref.current!.toggleHighContrast()}>
-            Contrast
-          </button>
-          <button onClick={() => ref.current!.closeFlyout()}>
-            Close flyout
-          </button>
-        </div>
-        <div style={EditorWithToolbarStyles.actions}>
-          <button
-            onClick={async () => {
-              const result = await ref.current!.renderBlocks({
-                ts: 'basic.showNumber(42)',
-              });
-              const img = document.body.appendChild(
-                document.createElement('img')
-              );
-              img.src = result!;
-            }}
-          >
-            Render blocks
-          </button>
-          <button
-            onClick={async () => {
-              const result = await ref.current!.renderPython({
-                ts: 'basic.showNumber(42)',
-              });
-              console.log(result);
-            }}
-          >
-            Render Python
-          </button>
-          <button
-            onClick={async () => {
-              const result = await ref.current!.renderXml({
-                xml: defaultMakeCodeProject.text!['main.blocks']!,
-              });
-              const img = document.body.appendChild(
-                document.createElement('img')
-              );
-              img.src = result!;
-            }}
-          >
-            Render XML
-          </button>
-          <button
-            onClick={async () => {
-              const result = await ref.current!.renderByBlockId({
-                blockId: 'basic_show_icon',
-              });
-              const img = document.body.appendChild(
-                document.createElement('img')
-              );
-              img.src = result!;
-            }}
-          >
-            Render by block id
-          </button>
-        </div>
-        <div style={EditorWithToolbarStyles.actions}>
-          <label>
-            File to import: <input type="file" id="importFile"></input>
-          </label>
-          <button
-            onClick={async () => {
-              const importFile = document.querySelector(
-                '#importFile'
-              ) as HTMLInputElement;
-              const file = importFile.files?.item(0);
-              if (file) {
-                const data = await file.arrayBuffer();
-                const text = new TextDecoder().decode(data);
-                ref.current!.importFile({
-                  filename: file.name,
-                  parts: [text],
-                });
-              }
-            }}
-          >
-            Import file
-          </button>
-        </div>
-        <div style={EditorWithToolbarStyles.actions}>
-          <button onClick={() => ref.current!.startSimulator()}>
-            Start simulator
-          </button>
-          <button onClick={() => ref.current!.stopSimulator()}>
-            Stop simulator
-          </button>
-          <button onClick={() => ref.current!.hideSimulator()}>
-            Hide simulator
-          </button>
-          <button
-            onClick={() => {
-              ref.current!.setSimulatorFullScreen(true);
-            }}
-          >
-            Set simulator full screen
-          </button>
-        </div>
-      </div>
+      <MakeCodeToolbar savedProjects={savedProjects} driver={ref} />
       <MakeCodeFrame
         ref={ref}
         controller={1}
